Updated render side scripts
Added function to detect if a port is connected or not.
When port is not connected, function findPort is called every 2 secconds.
If port is connected function portFound is called once and never again. If port is disconnected find port is called every 2 secconds again. 
And at the and if port is connected again portFound is called once again.